requirements.txt
Flask
google-ads
PyYAML
gunicorn